Update transition value and springValue on props change
