+++ src/index.ts
@@ -1,4 +1,3 @@
 export type { SpringOptions } from './spring';
 export type { Stage, Values, Options } from './useSpringTransition';
 export { STAGES, useSpringTransition } from './useSpringTransition';
-export { useSpringTransitions } from './useSpringTransitions';
